backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
redis==5.2.1
neo4j==5.27.0
qdrant-client==1.12.1
httpx==0.28.1
torch==2.6.0
torchvision==0.21.0
Pillow==12.2.0
openai-whisper==20240930
gTTS==2.5.4
numpy==2.2.1
scikit-learn==1.6.1
pdfplumber==0.11.4
pytesseract==0.3.13
python-dotenv==1.0.1
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
structlog==24.4.0
pytest==8.3.4
pytest-asyncio==0.25.0
aiofiles==24.1.0
websockets==14.1
